expr: Adds a new LIKE operator that is more efficient in the common case.
The existing LIKE implementation converts the pattern to a regular expression and uses Rust's Regex library to perform the matching. This works, but regular expressions tend to be inefficient for simple string searches.

This PR introduces a second implementation that uses std::str methods to perform the searching. The compile step is ~100x faster, and common cases are around 2-3x faster. (Benchmarks below.)

Not all patterns are faster. The regular expressions maintain consistent performance when the number of wildcards increase, but the new implementation gets more expensive as the complexity of the pattern increases.

To demonstrate this, I have included an adversarial pattern in the benchmarks. (`like_search_expensive`) This is an extreme example that is unlikely to occur in practice. Regex is able to reject the match in 48 ns due because it recognizes early that the last character `?` does not appear in the pattern. The new implmenetaiton, however, runs through O(N^2) substring searches before rejecting the match.

```
Bechmark	             Regex (ns)	  New (ns)	% Change
like_compile	          18804.00	   183.29	    -99%
like_search_literal	      388.76	   128.92	    -67%
like_search_starts_with 	919.31	   133.00	    -86%
like_search_contains	   1224.70	   633.23	    -48%
like_search_contains2	   1223.30	   729.70	    -40%
like_search_contains3	   1222.40	   757.45	    -38%
like_search_ends_with	   1299.20	   717.31	    -45%
like_search_expensive	     48.02	925480.00	 +1927020% (!!!)
```

A future improvement might be to fall back to the Regex implementation if the pattern appears to be too complex. This would likely reduce the risk of adversarial patterns.