persist: react immediately to state changes in the same process
This came out of some skunkworks explorations around pushing persist
state. That's still ongoing, but this seemed to be a minor win and
uncontroversial enough that it was worth pulling it out.

Now that we have at most one copy of State in a process for a given
shard, there's an easy place to hook into to notify anyone else in the
same process who might be waiting for a new version of the state
(primarily Listens).