[persist/expr] Fix a bug when handling associative variadic functions (#19920)
This bug is an bad interaction between two recently-added features:
- In `ColumnSpecs::variadic`, we check if a function is associative; if
so, we evaluate it as a series of nested calls.
- To support lazy variadic funcs - which means that arguments to
functions can be `Err`s - we evaluate functions by passing arguments as
literals instead of using column references. To make this efficient, we
allocate a single expression and update the literals in place, checking
that we don't pass a type we didn't expect up front.

In brief, the associative optimization thinks that some nested calls may
return null, but the type in our placeholder does not expect null. I've
left inline comments that try to walk through the issue.

This PR includes some test improvements, as well as two fixes:
- The first fix fixes the bug in a targeted way by filtering the result
to only results that match the output type. Since the output type does
not include `null`, we'll never end up trying to set `null` in the
placeholder.
- The second replaces our special logic with calls to the regular
`variadic` function. (Since every call to `variadic` already checks the
type at the end, this is equivalent.)

Both approaches pass tests, but I find the second one leaves the code in
a more readable state.